+++ starkinfra/__init__.py
@@ -47,8 +47,17 @@ from .issuingbalance.__issuingbalance import IssuingBalance
 from . import creditnote
 from .creditnote.__creditnote import CreditNote
 
+from . import creditsigner
+from .creditsigner.__creditsigner import CreditSigner
+
 from . import creditpreview
 from .creditpreview.__creditpreview import CreditPreview
+
+from . import individualidentity
+from .individualidentity.__individualidentity import IndividualIdentity
+
+from . import individualdocument
+from .individualdocument.__individualdocument import IndividualDocument
 
 from . import dynamicbrcode
 from .dynamicbrcode.__dynamicbrcode import DynamicBrcode