+++ starkinfra/individualdocument/__init__.py
@@ -1,0 +1,3 @@
+from . import log
+from .log.__log import Log
+from .__individualdocument import create, get, query, page